hide 'clear your answers and try again' unless allow_redo is set. PMT #88601
There's still some weirdness to be aware of here. Specifically, if there are more than one
blocks on the page that take submissions and they are set to a mix of allow/disallow redo
it will allow it if Any of them allow it. I don't know what the use-case for a mixed
permission page like that is, but I'm sure an ET will *need* one at some point