Add csrf_token to pagetree admin forms
The csrf_token template tag was only present on one of the forms here..
since ccnmtldjango has csrf checking on by default, I think we include
the csrf_tokens here.

I tested the csrf_token templatetag when the CsrfViewMiddleware is
disabled, and the forms still work.. so if you need to turn it off for
some reason it won't require a change to pagetree.