Fix bug where cached value wasn't used when value was 0
When debugging the `is_last_child` cache, I saw the value
of cache.get() here is either 1, 0, or None if the cached
value isn't present. Currently, only values of 1 are getting
used, because if the value is 0, the `if v:` to check if
the cache is present fails. I think we need to be more
specific here and do: `if v is not None:`

This change improved performance a little bit in my local testing
with memcached, but it's still not great: Calling the cache for
each node in the tree takes time, and it would be great to find
a way to interact with the cache just once for the pagetree edit
page.

refs #63