migration to delete TestBlock table if it exists
At one point in the past, the `TestBlock` model was visible to the
migration system and a bunch of our apps have a stray
`pagetree_testblock` table in the database. When `manage.py migrate` is
run against those databases, we get:

```
  Your models have changes that are not yet reflected in a migration, and so won't be applied.
  Run 'manage.py makemigrations' to make new migrations, and then re-run 'manage.py migrate' to apply them.
```

Running `manage.py makemigrations` will create a pagetree migration that
deletes the table.

Unfortunately, if that gets checked in, it breaks the apps that managed
to avoid getting that `pagetree_testblock` table created in the first
place (it tries to delete a table that doesn't exist and that fails).

Even more importantly, it breaks the unit tests since the in-memory
database doesn't have that table.

This commit adds a migration which does the
`migrations.DeleteModel(name="TestBlock")`, but wraps it in a try/catch
so it's harmless if the table doesn't actually exist.